+++ tests/test_expense.py
@@ -1,9 +1,7 @@
-import os
-from sqlalchemy.sql.elements import not_
-from citywok_ms.employee.models import Employee
 import datetime
 import html
 import io
+import os
 
 import pytest
 from citywok_ms import db
@@ -14,23 +12,20 @@ from citywok_ms.expense.models import (
     NonLaborExpense,
     SalaryPayment,
 )
-from flask import request, url_for
-from sqlalchemy.orm.util import with_polymorphic
-from wtforms.fields.simple import FileField, HiddenField, SubmitField
-from citywok_ms.supplier.models import Supplier
 from citywok_ms.file.models import ExpenseFile, SalaryPaymentFile
+from flask import request, url_for
+from sqlalchemy.sql.elements import not_
+from wtforms.fields.simple import BooleanField, FileField, HiddenField, SubmitField
 
 
 @pytest.mark.role("admin")
-def test_index_get(client, user, expenses):
+def test_index_get(client, user, expenses, today):
     response = client.get(url_for("expense.index"), follow_redirects=True)
     data = response.data.decode()
 
     # state code
     assert response.status_code == 200
-    assert request.url.endswith(
-        url_for("expense.index", date_str=datetime.date.today())
-    )
+    assert request.url.endswith(url_for("expense.index", date_str=today))
 
     for txt in (
         "Expenses",
@@ -54,15 +49,14 @@ def test_index_get(client, user, expenses):
 
 
 @pytest.mark.role("admin")
-def test_index_post(client, user):
-    date = datetime.datetime.today() - datetime.timedelta(days=1)
+def test_index_post(client, user, yesterday, today):
     response = client.post(
-        url_for("expense.index", date_str=datetime.date.today()),
-        data={"date": date.date()},
+        url_for("expense.index", date_str=today),
+        data={"date": yesterday},
         follow_redirects=True,
     )
     assert response.status_code == 200
-    assert request.url.endswith(url_for("expense.index", date_str=date.date()))
+    assert request.url.endswith(url_for("expense.index", date_str=yesterday))
 
 
 @pytest.mark.role("admin")
@@ -77,7 +71,7 @@ def test_new_non_labor_get(client, user, supplier):
     assert "New Non-Labor Expense" in data
 
     for field in NonLaborExpenseForm()._fields.values():
-        if isinstance(field, (HiddenField, SubmitField)):
+        if isinstance(field, (HiddenField, SubmitField, BooleanField)):
             continue
         assert field.id in data
     assert "Add" in data
@@ -86,10 +80,9 @@ def test_new_non_labor_get(client, user, supplier):
 
 
 @pytest.mark.role("admin")
-def test_new_non_labor_post_valid(client, user, supplier, image):
-    today = datetime.datetime.today()
+def test_new_non_labor_post_valid(client, user, supplier, image, today):
     request_data = {
-        "date": today.date(),
+        "date": today,
         "category": "operation:water",
         "supplier": 1,
         "value-cash": 5,
@@ -106,13 +99,13 @@ def test_new_non_labor_post_valid(client, user, supplier, image):
     # state code
     assert response.status_code == 200
     # url after request
-    assert request.url.endswith(url_for("expense.index", date_str=today.date()))
+    assert request.url.endswith(url_for("expense.index", date_str=today))
 
     # database data
     assert db.session.query(Expense).count() == 1
     expense = db.session.query(NonLaborExpense).first()
     assert expense.total == 20
-    assert expense.date == today.date()
+    assert expense.date == today
     assert expense.supplier_id == 1
     assert expense.category == "operation:water"
     assert len(expense.files) == 1
@@ -141,6 +134,29 @@ def test_new_non_labor_post_invalid(client, user):
 
 
 @pytest.mark.role("admin")
+def test_new_non_labor_post_from_pos(client, user, today, image):
+    request_data = {
+        "from_pos": True,
+        "value-cash": 5,
+        "value-card": 5,
+    }
+    response = client.post(
+        url_for("expense.new_non_labor"), data=request_data, follow_redirects=True
+    )
+    data = response.data.decode()
+
+    # state code
+    assert response.status_code == 200
+    # url after request
+    assert request.url.endswith(url_for("expense.new_non_labor"))
+
+    # database data
+    assert db.session.query(Expense).count() == 0
+
+    assert "Expense from POS must be payed with cash." in html.unescape(data)
+
+
+@pytest.mark.role("admin")
 def test_new_labor_get(client, user, employee):
     response = client.get(url_for("expense.new_labor", employee_id=1))
     data = response.data.decode()
@@ -152,7 +168,7 @@ def test_new_labor_get(client, user, employee):
     assert "New Labor Expense" in data
 
     for field in NonLaborExpenseForm()._fields.values():
-        if isinstance(field, (HiddenField, SubmitField)):
+        if isinstance(field, (HiddenField, SubmitField, BooleanField)):
             continue
         assert field.id in data
     assert "Add" in data
@@ -161,10 +177,9 @@ def test_new_labor_get(client, user, employee):
 
 
 @pytest.mark.role("admin")
-def test_new_labor_post_valid(client, user, employee, image):
-    today = datetime.datetime.today()
+def test_new_labor_post_valid(client, user, employee, image, today):
     request_data = {
-        "date": today.date(),
+        "date": today,
         "category": "labor:advance",
         "employee": 1,
         "value-cash": 5,
@@ -181,13 +196,13 @@ def test_new_labor_post_valid(client, user, employee, image):
     # state code
     assert response.status_code == 200
     # url after request
-    assert request.url.endswith(url_for("expense.index", date_str=today.date()))
+    assert request.url.endswith(url_for("expense.index", date_str=today))
 
     # database data
     assert db.session.query(Expense).count() == 1
     expense = db.session.query(LaborExpense).first()
     assert expense.total == 20
-    assert expense.date == today.date()
+    assert expense.date == today
     assert expense.employee_id == 1
     assert expense.category == "labor:advance"
     assert len(expense.files) == 1
@@ -214,6 +229,33 @@ def test_new_labor_post_invalid(client, user):
 
 
 @pytest.mark.role("admin")
+def test_new_labor_post_from_pos(client, user, today, image):
+    request_data = {
+        "date": today,
+        "category": "labor:advance",
+        "employee": 1,
+        "from_pos": True,
+        "value-cash": 5,
+        "value-card": 5,
+        "files": (image, "test.jpg"),
+    }
+    response = client.post(
+        url_for("expense.new_labor"), data=request_data, follow_redirects=True
+    )
+    data = response.data.decode()
+
+    # state code
+    assert response.status_code == 200
+    # url after request
+    assert request.url.endswith(url_for("expense.new_labor"))
+
+    # database data
+    assert db.session.query(Expense).count() == 0
+
+    assert "Expense from POS must be payed with cash." in html.unescape(data)
+
+
+@pytest.mark.role("admin")
 def test_new_order_payment_get(client, user, supplier):
     response = client.get(url_for("expense.new_order_payment", supplier_id=1))
     data = response.data.decode()
@@ -225,7 +267,7 @@ def test_new_order_payment_get(client, user, supplier):
     assert "New Orders Payment" in data
 
     for field in NonLaborExpenseForm()._fields.values():
-        if isinstance(field, (HiddenField, SubmitField)):
+        if isinstance(field, (HiddenField, SubmitField, BooleanField)):
             continue
         assert field.id in data
     assert "Add" in data
@@ -234,10 +276,9 @@ def test_new_order_payment_get(client, user, supplier):
 
 
 @pytest.mark.role("admin")
-def test_new_order_payment_post_valid(client, user, supplier, image, order):
-    today = datetime.datetime.today()
+def test_new_order_payment_post_valid(client, user, supplier, image, order, today):
     request_data = {
-        "date": today.date(),
+        "date": today,
         "category": "material:meat",
         "supplier": 1,
         "orders": 1,
@@ -256,13 +297,13 @@ def test_new_order_payment_post_valid(client, user, supplier, image, order):
     # state code
     assert response.status_code == 200
     # url after request
-    assert request.url.endswith(url_for("expense.index", date_str=today.date()))
+    assert request.url.endswith(url_for("expense.index", date_str=today))
 
     # database data
     assert db.session.query(Expense).count() == 1
     expense = db.session.query(NonLaborExpense).first()
     assert expense.total == 20
-    assert expense.date == today.date()
+    assert expense.date == today
     assert expense.supplier_id == 1
     assert expense.orders[0].id == 1
     assert expense.category == "material:meat"
@@ -313,7 +354,10 @@ def test_new_salary_get(client, user, employee):
     assert "Payment" in data
 
     for field in NonLaborExpenseForm()._fields.values():
-        if isinstance(field, (HiddenField, SubmitField)) or field.id == "category":
+        if (
+            isinstance(field, (HiddenField, SubmitField, BooleanField))
+            or field.id == "category"
+        ):
             continue
         assert field.id in data
     assert "Add" in data
@@ -323,11 +367,9 @@ def test_new_salary_get(client, user, employee):
 
 
 @pytest.mark.role("admin")
-def test_new_salary_post_valid(client, user, employee, image):
-    today = datetime.datetime.today()
-    month_str = today.strftime("%Y-%m")
+def test_new_salary_post_valid(client, user, employee, image, today, current_month):
     request_data = {
-        "date": today.date(),
+        "date": today,
         "value-cash": 5,
         "value-card": 5,
         "value-check": 5,
@@ -335,7 +377,7 @@ def test_new_salary_post_valid(client, user, employee, image):
         "files": (image, "test.jpg"),
     }
     response = client.post(
-        url_for("expense.new_salary", month_str=month_str, employee_id=1),
+        url_for("expense.new_salary", month_str=current_month, employee_id=1),
         data=request_data,
         follow_redirects=True,
     )
@@ -344,17 +386,19 @@ def test_new_salary_post_valid(client, user, employee, image):
     # state code
     assert response.status_code == 200
     # url after request
-    assert request.url.endswith(url_for("expense.salary_index", month_str=month_str))
+    assert request.url.endswith(
+        url_for("expense.salary_index", month_str=current_month)
+    )
 
     # database data
     assert db.session.query(Expense).count() == 1
     assert db.session.query(SalaryPayment).count() == 1
     expense = db.session.query(LaborExpense).first()
     assert expense.total == 20
-    assert expense.date == today.date()
+    assert expense.date == today
     assert expense.employee_id == 1
     assert expense.category == "labor:salary"
-    assert expense.month_id == today.replace(day=1).date()
+    assert expense.month_id == today.replace(day=1)
     assert len(expense.files) == 1
 
     # flash messege
@@ -362,10 +406,9 @@ def test_new_salary_post_valid(client, user, employee, image):
 
 
 @pytest.mark.role("admin")
-def test_new_salary_post_invalid(client, user, employee):
-    month_str = datetime.datetime.today().strftime("%Y-%m")
+def test_new_salary_post_invalid(client, user, employee, current_month):
     response = client.post(
-        url_for("expense.new_salary", month_str=month_str, employee_id=1),
+        url_for("expense.new_salary", month_str=current_month, employee_id=1),
         follow_redirects=True,
     )
     data = response.data.decode()
@@ -374,7 +417,7 @@ def test_new_salary_post_invalid(client, user, employee):
     assert response.status_code == 200
     # url after request
     assert request.url.endswith(
-        url_for("expense.new_salary", month_str=month_str, employee_id=1)
+        url_for("expense.new_salary", month_str=current_month, employee_id=1)
     )
 
     # database data
@@ -385,10 +428,9 @@ def test_new_salary_post_invalid(client, user, employee):
 
 
 @pytest.mark.role("admin")
-def test_new_salary_payed(client, user, expenses):
-    month_str = datetime.datetime.today().strftime("%Y-%m")
+def test_new_salary_payed(client, user, expenses, current_month):
     response = client.get(
-        url_for("expense.new_salary", month_str=month_str, employee_id=1),
+        url_for("expense.new_salary", month_str=current_month, employee_id=1),
         follow_redirects=True,
     )
     data = response.data.decode()
@@ -397,20 +439,20 @@ def test_new_salary_payed(client, user, expenses):
     assert response.status_code == 200
 
     assert "Employee already payed at the given month." in data
-    assert request.url.endswith(url_for("expense.salary_index", month_str=month_str))
+    assert request.url.endswith(
+        url_for("expense.salary_index", month_str=current_month)
+    )
 
 
 @pytest.mark.role("admin")
-def test_salary_index_get(client, user, expenses):
+def test_salary_index_get(client, user, expenses, current_month):
     response = client.get(url_for("expense.salary_index"), follow_redirects=True)
     data = response.data.decode()
 
     # state code
     assert response.status_code == 200
     assert request.url.endswith(
-        url_for(
-            "expense.salary_index", month_str=datetime.date.today().strftime("%Y-%m")
-        )
+        url_for("expense.salary_index", month_str=current_month)
     )
 
     for txt in (
@@ -426,17 +468,14 @@ def test_salary_index_get(client, user, expenses):
 
 
 @pytest.mark.role("admin")
-def test_salary_index_post(client, user):
-    month = (
-        datetime.datetime.today().replace(day=1) - datetime.timedelta(days=1)
-    ).strftime("%Y-%m")
+def test_salary_index_post(client, user, last_month):
     response = client.post(
-        url_for("expense.salary_index", month_str=month),
-        data={"month": month},
+        url_for("expense.salary_index", month_str=last_month),
+        data={"month": last_month},
         follow_redirects=True,
     )
     assert response.status_code == 200
-    assert request.url.endswith(url_for("expense.salary_index", month_str=month))
+    assert request.url.endswith(url_for("expense.salary_index", month_str=last_month))
 
 
 @pytest.mark.role("admin")
@@ -488,7 +527,7 @@ def test_update_non_labor_get(client, user, expenses):
     assert "Update Non-Labor Expense" in data
 
     for field in NonLaborExpenseForm()._fields.values():
-        if isinstance(field, (HiddenField, SubmitField, FileField)):
+        if isinstance(field, (HiddenField, SubmitField, BooleanField, FileField)):
             continue
         assert field.id in data
     assert "Update" in data
@@ -497,11 +536,10 @@ def test_update_non_labor_get(client, user, expenses):
 
 
 @pytest.mark.role("admin")
-def test_update_non_labor_post_valid(client, user, expenses, image):
+def test_update_non_labor_post_valid(client, user, expenses, image, today):
     expense = NonLaborExpense.query.filter(not_(NonLaborExpense.orders)).first()
-    today = datetime.datetime.today()
     request_data = {
-        "date": today.date(),
+        "date": today,
         "category": "operation:gas",
         "supplier": 1,
         "value-cash": 10,
@@ -525,7 +563,7 @@ def test_update_non_labor_post_valid(client, user, expenses, image):
     assert db.session.query(Expense).count() == 4
     expense = NonLaborExpense.query.filter(not_(NonLaborExpense.orders)).first()
     assert expense.total == 40
-    assert expense.date == today.date()
+    assert expense.date == today
     assert expense.supplier_id == 1
     assert expense.category == "operation:gas"
 
@@ -576,7 +614,7 @@ def test_update_labor_get(client, user, expenses):
     assert "Update Labor Expense" in data
 
     for field in NonLaborExpenseForm()._fields.values():
-        if isinstance(field, (HiddenField, SubmitField, FileField)):
+        if isinstance(field, (HiddenField, SubmitField, BooleanField, FileField)):
             continue
         assert field.id in data
     assert "Update" in data
@@ -585,11 +623,10 @@ def test_update_labor_get(client, user, expenses):
 
 
 @pytest.mark.role("admin")
-def test_update_labor_post_valid(client, user, expenses, image):
+def test_update_labor_post_valid(client, user, expenses, image, today):
     expense = LaborExpense.query.filter(not_(LaborExpense.month.has())).first()
-    today = datetime.datetime.today()
     request_data = {
-        "date": today.date(),
+        "date": today,
         "category": "labor:bonus",
         "employee": 1,
         "value-cash": 10,
@@ -613,7 +650,7 @@ def test_update_labor_post_valid(client, user, expenses, image):
     assert db.session.query(Expense).count() == 4
     expense = LaborExpense.query.filter(not_(LaborExpense.month.has())).first()
     assert expense.total == 40
-    assert expense.date == today.date()
+    assert expense.date == today
     assert expense.employee_id == 1
     assert expense.category == "labor:bonus"
 
@@ -664,7 +701,7 @@ def test_update_order_payment_get(client, user, expenses):
     assert "Update Orders Payment" in data
 
     for field in NonLaborExpenseForm()._fields.values():
-        if isinstance(field, (HiddenField, SubmitField, FileField)):
+        if isinstance(field, (HiddenField, SubmitField, BooleanField, FileField)):
             continue
         assert field.id in data
     assert "Update" in data
@@ -673,11 +710,10 @@ def test_update_order_payment_get(client, user, expenses):
 
 
 @pytest.mark.role("admin")
-def test_update_order_payment_post_valid(client, user, expenses, image):
+def test_update_order_payment_post_valid(client, user, expenses, image, today):
     expense = NonLaborExpense.query.filter(NonLaborExpense.orders).first()
-    today = datetime.datetime.today()
     request_data = {
-        "date": today.date(),
+        "date": today,
         "category": "operation:gas",
         "supplier": 1,
         "value-cash": 10,
@@ -703,7 +739,7 @@ def test_update_order_payment_post_valid(client, user, expenses, image):
     assert db.session.query(Expense).count() == 4
     expense = NonLaborExpense.query.filter(NonLaborExpense.orders).first()
     assert expense.total == 40
-    assert expense.date == today.date()
+    assert expense.date == today
     assert expense.supplier_id == 1
     assert len(expense.orders) == 1
     assert expense.category == "operation:gas"
@@ -756,7 +792,7 @@ def test_update_salary_get(client, user, expenses):
 
     for field in NonLaborExpenseForm()._fields.values():
         if (
-            isinstance(field, (HiddenField, SubmitField, FileField))
+            isinstance(field, (HiddenField, SubmitField, BooleanField, FileField))
             or field.id == "category"
         ):
             continue
@@ -767,11 +803,10 @@ def test_update_salary_get(client, user, expenses):
 
 
 @pytest.mark.role("admin")
-def test_update_salary_post_valid(client, user, expenses, image):
+def test_update_salary_post_valid(client, user, expenses, image, today):
     expense = LaborExpense.query.filter(LaborExpense.month.has()).first()
-    today = datetime.datetime.today()
     request_data = {
-        "date": today.date(),
+        "date": today,
         "value-cash": 10,
         "value-card": 10,
         "value-check": 10,
@@ -793,7 +828,7 @@ def test_update_salary_post_valid(client, user, expenses, image):
     assert db.session.query(Expense).count() == 4
     expense = LaborExpense.query.filter(LaborExpense.month.has()).first()
     assert expense.total == 40
-    assert expense.date == today.date()
+    assert expense.date == today
     assert expense.employee_id == 1
 
     # flash messege
@@ -881,20 +916,21 @@ def test_upload_post_invalid_empty(client, user, expenses):
 
 
 @pytest.mark.role("admin")
-def test_salary_upload_post_valid(client, user, expenses, image):
-    month = datetime.datetime.today().strftime("%Y-%m")
+def test_salary_upload_post_valid(client, user, expenses, image, current_month):
     request_data = {
         "file": (image, "test.jpg"),
     }
     response = client.post(
-        url_for("expense.salary_upload", month_str=month),
+        url_for("expense.salary_upload", month_str=current_month),
         data=request_data,
         content_type="multipart/form-data",
         follow_redirects=True,
     )
     data = response.data.decode()
     assert response.status_code == 200
-    assert request.url.endswith(url_for("expense.salary_index", month_str=month))
+    assert request.url.endswith(
+        url_for("expense.salary_index", month_str=current_month)
+    )
     assert 'File "test.jpg" has been uploaded.' in html.unescape(data)
     assert db.session.query(SalaryPaymentFile).count() == 1
     f = db.session.query(SalaryPaymentFile).get(1)
@@ -903,13 +939,12 @@ def test_salary_upload_post_valid(client, user, expenses, image):
 
 
 @pytest.mark.role("admin")
-def test_salary_upload_post_invalid_format(client, user, expenses):
-    month = datetime.datetime.today().strftime("%Y-%m")
+def test_salary_upload_post_invalid_format(client, user, expenses, current_month):
     request_data = {
         "file": (io.BytesIO(b"test"), "test.exe"),
     }
     response = client.post(
-        url_for("expense.salary_upload", month_str=month),
+        url_for("expense.salary_upload", month_str=current_month),
         data=request_data,
         content_type="multipart/form-data",
         follow_redirects=True,
@@ -917,15 +952,16 @@ def test_salary_upload_post_invalid_format(client, user, expenses):
     data = response.data.decode()
 
     assert response.status_code == 200
-    assert request.url.endswith(url_for("expense.salary_index", month_str=month))
+    assert request.url.endswith(
+        url_for("expense.salary_index", month_str=current_month)
+    )
     assert 'Invalid file format ".exe".' in html.unescape(data)
 
 
 @pytest.mark.role("admin")
-def test_salary_upload_post_invalid_empty(client, user, expenses):
-    month = datetime.datetime.today().strftime("%Y-%m")
+def test_salary_upload_post_invalid_empty(client, user, expenses, current_month):
     response = client.post(
-        url_for("expense.salary_upload", month_str=month),
+        url_for("expense.salary_upload", month_str=current_month),
         data={},
         content_type="multipart/form-data",
         follow_redirects=True,
@@ -933,7 +969,9 @@ def test_salary_upload_post_invalid_empty(client, user, expenses):
     data = response.data.decode()
 
     assert response.status_code == 200
-    assert request.url.endswith(url_for("expense.salary_index", month_str=month))
+    assert request.url.endswith(
+        url_for("expense.salary_index", month_str=current_month)
+    )
     assert "No file has been uploaded." in html.unescape(data)